// client/dist/widgets/common/list/src/config.ts
export interface Suggestion {
  suggestionHtml: string | Element
  suggestion: string
}
// ...
interface WidgetStyle {
  id: string
}
// ...
export enum PageStyle {
  Scroll = 'SCROLL',
  MultiPage = 'MULTIPAGE'
}

export enum AlignType {
  Start = 'FLEX-START',
  Center = 'CENTER',
  End = 'FLEX-END'
}

export enum DirectionType {
  Horizon = 'HORIZON',
  Vertical = 'VERTICAL'
}

export enum ListLayoutType {
  Row = 'ROW',
  Column = 'COLUMN',
  GRID = 'GRID'
}

export enum PageTransitonType {
  Glide = 'GLIDE',
  Fade = 'FADE',
  Float = 'FLOAT'
}

export enum HoverType {
  Hover0 = 'HOVER0',
  Hover1 = 'HOVER1',
  Hover2 = 'HOVER2',
  Hover3 = 'HOVER3'
}

export enum SelectedStyle {
  Style0 = 'STYLE0',
  Style1 = 'STYLE1',
  Style2 = 'STYLE2',
  Style3 = 'STYLE3'
}

export enum ItemStyle {
  Style0 = 'STYLE0',
  Style1 = 'STYLE1',
  Style2 = 'STYLE2',
  Style3 = 'STYLE3',
  Style4 = 'STYLE4',
  Style5 = 'STYLE5',
  Style6 = 'STYLE6',
  Style7 = 'STYLE7',
  Style8 = 'STYLE8',
  Style9 = 'STYLE9'
}
// ...
export interface CardConfig {
  backgroundStyle?: CardBackgroundStyle
  enable?: boolean
  selectionMode?: SelectionModeType
  cardSize?: DeviceCardSize
  listLayout?: ListLayout
}

export type IMCardConfig = ImmutableObject<CardConfig>

export interface Config {
  pageTransition?: PageTransitonType
  hoverType?: HoverType
  selectedStyle?: SelectedStyle
  differentOddEven?: boolean
  itemStyle?: ItemStyle
  isCheckEmptyTemplate?: boolean
  isItemStyleConfirm?: boolean
  direction?: DirectionType
  alignType?: AlignType
  space?: number
  horizontalSpace?: number
  verticalSpace?: number
  itemsPerPage?: number
  pageStyle?: PageStyle
  scrollBarOpen?: boolean
  navigatorOpen?: boolean
  scrollStep?: number
  style?: ImmutableObject<WidgetStyle>
  isInitialed?: boolean
  lockItemRatio?: boolean
  showSelectedOnlyOpen?: boolean
  showClearSelected?: boolean
  gridAlignment?: TextAlignValue

  // link
  linkParam?: IMLinkParam

  // search
  searchOpen?: boolean
  searchFields?: string
  searchExact?: boolean
  // filter
  filterOpen?: boolean
  filter?: SqlExpression
  // sort
  sortOpen?: boolean
  sorts?: SortSettingOption[]

  // card background
  cardConfigs?: ImmutableObject<{ [status: string]: CardConfig }>
  searchHint?: string

  isShowAutoRefresh?: boolean
  noDataMessage?: string
  layoutType?: ListLayoutType
  keepAspectRatio?: boolean
  gridItemSizeRatio?: number
}

export type IMConfig = ImmutableObject<Config>